v6: factor out matching logic
This simplifies the code a little and prepares for the possibility
of preprocessing the match code for faster matching.
